requirements.txt
tweepy>=4.14.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
pytz==2024.1
deep-translator==1.11.4
APScheduler==3.10.4
beautifulsoup4==4.12.3
yfinance==0.2.33

# Improvements: Charts, Caching, AI
matplotlib>=3.9.0
pillow>=10.4.0
groq>=0.4.0
